refactor: split web/reports.py into assembly, document and label layers
`mureo/web/reports.py` was 1,070 lines holding twenty-six functions across
eight concerns: platform naming, the tolerant STATE.json read, conflict
detection, freshness, the period vocabulary, `not_collected` notes, wire
sanitisation, and the summary assembly that uses all of them.

Three files now, cut along the direction the data already flows:

  mureo/web/report_labels.py (185 lines)
    Platform key -> the name an operator reads. Three cases, and the third
    carries the weight: an unrecognisable key is returned UNCHANGED, which
    is precisely how `CONFLICT_UNRECOGNIZED_KEY` is defined. A prettier
    fallback invented here would silently switch off a conflict finding a
    module away, so the two now sit one import apart and say so.

  mureo/web/report_document.py (624 lines)
    Everything asked OF one stored document: the tolerant read, the two
    conflict findings (#533), freshness (#535), which windows the document
    carries, its `not_collected` notes (#638), and the canonical-key filter
    that decides what may reach the wire.

  mureo/web/reports.py (386 lines)
    The assembly: resolve the store, walk the document once, shape the
    payload. Plus the facade every existing importer already targets.

Import direction is one-way — reports -> report_document -> report_labels —
so there is no cycle to reason about.

Why this file could be cut where mcp/server.py could not
--------------------------------------------------------
The two mechanisms that pinned `server.py` in place are both absent here,
measured rather than assumed:

  - Nothing reloads `mureo.web.reports`. `importlib.reload` appears nowhere
    against it, so no module-level statement has to stay put to be re-run.
  - Nothing monkey-patches a name on it. All twenty runtime-context patches
    in the suite target `mureo.web.report_clients.get_runtime_context`, a
    module that was already split out. The one clock patch goes to
    `mureo.core.clock.server_now`, and this module reaches it as a module
    attribute (`clock.server_now()`), which survives any move.
  - No test reads a `mureo/**.py` as source text, so there are no static
    grep pins of the kind that block the `dashboard.js` half of #678.

No behaviour change, no API change
----------------------------------
Every name is re-exported from `mureo.web.reports`: the nine in `__all__`,
and the four private tables the suite reads off it directly
(`_BUILTIN_DISPLAY_NAMES`, `_OFFICIAL_BRIDGE_DISPLAY_NAMES`, `_PERIOD_ORDER`,
`_PERIOD_LENGTH_DAYS`). `from mureo.web.reports import platform_display_name`
and `mureo.web.reports._BUILTIN_DISPLAY_NAMES` resolve exactly as before.

Asserted mechanically rather than by inspection: all twenty-six top-level
definitions are present across the three files with identical `ast.dump`
output — nothing missing, nothing added, no body changed. Of the thirty-one
module-level statements, twenty-nine survive verbatim; the two that do not
are the module docstring (extended, and the original text is a verbatim
prefix of the new one) and the `if TYPE_CHECKING:` block, split across the
three files with its name set preserved.

Ten docstring cross-references in `mureo/` are re-pointed at the module the
function moved to. `mureo.web.reports._platform_row` is deliberately left
alone — that one stayed.

Test changes are two prose strings and nothing else: an assertion MESSAGE in
`tests/test_web_reports_conflicts.py` and a docstring in `tests/test_state.py`,
both naming a moved helper. No assertion, no fixture and no import changed.

Refs #678
